# openapi-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EvoNEST API",
    "version": "2.0.0",
    "description": "\n# EvoNEST General Documentation\n\n**The EvoNEST User Guide and the EvoNEST Developer Guide are found here**: https://daniele-liprandi.github.io/EvoNEST-backbone/\n\n# EvoNEST API Documentation\n\nThe EvoNEST API allows users to interact with the MongoDB databases generated in their NEST, plus it gives access to utility functions useful in ecological, evolutionary and biological research.\n\n## Features\n\n- **Samples Management**: Create, update, and retrieve biological samples with taxonomic information\n- **Traits Measurement**: Record and analyze trait measurements with statistical calculations\n- **Experiments**: Manage measurements, documents and raw data\n- **File Storage**: Upload and manage research files with metadata\n- **User Management**: Handle user accounts and permissions\n- **Utilities**: Geocoding, species image search, and data processing tools\n\n## Authentication\n\nMost endpoints require proper authentication and database access permissions.\n\n## Data Format\n\nAll endpoints return JSON data unless otherwise specified. Dates are in ISO 8601 format.\n      ",
    "contact": {
      "name": "Daniele Liprandi",
      "email": "daniele.liprandi@gmail.com",
      "url": "https://github.com/daniele-liprandi/EvoNEST-backbone"
    },
    "license": {
      "name": "GNU Affero General Public License v3.0",
...
    },
...
    "x-build-context": "GitHub Actions"
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "sessionAuth": []
    }
  ],
  "paths": {
...
    "/api/traits/analysis": {
      "post": {
        "summary": "Perform statistical analysis on trait data",
        "description": "Analyze trait measurements with statistical calculations (mean, stddev, min, max, median) grouped by various sample features",
        "tags": [
          "Traits"
        ],
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "traitType"
                ],
                "properties": {
                  "traitType": {
                    "type": "string",
                    "description": "Type of trait to analyze (e.g., stressAtBreak, diameter, toughness)"
                  },
                  "groupBy": {
                    "type": "string",
                    "default": "all",
                    "description": "Field to group results by (e.g., species, sex, sampleType)"
                  },
                  "filters": {
                    "type": "object",
                    "description": "Optional filters to apply to the data"
                  },
                  "unitConversion": {
                    "type": "boolean",
                    "default": true,
                    "description": "Whether to apply unit conversions"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Analysis completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "group": {
                            "type": "string"
                          },
                          "statistics": {
                            "type": "object",
                            "properties": {
                              "mean": {
                                "type": "number"
                              },
                              "stddev": {
                                "type": "number"
                              },
                              "min": {
                                "type": "number"
                              },
                              "max": {
                                "type": "number"
                              },
                              "median": {
                                "type": "number"
                              },
                              "count": {
                                "type": "number"
                              }
                            }
                          }
                        }
                      }
                    },
                    "processingTime": {
                      "type": "number",
                      "description": "Time taken for analysis in milliseconds"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Missing or invalid parameters"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  }
}